requirements.txt
tensorflow==2.1.0
tensorflow-hub==0.8.0
Flask==1.1.1
Pillow==7.0.0
matplotlib==3.1.3
opencv-python==4.2.0.32
requests==2.23.0